pyproject.toml
[project]
name = "skillberry-store"
version = "0.1.0"
description = "SkillBerry Store service"
authors = [{ name = "Eran Raichstein", email = "eranra@il.ibm.com" }, { name = "David Breitgand", email = "davidbr@il.ibm.com" }, { name = "Avi Weit", email = "weit@il.ibm.com" }]
dependencies = [
  "annotated-types>=0.7.0",
  "maturin>=1.8.4",
  "cargo",
  "anyio>=4.5.0",
  "asgiref>=3.8.1",
  "attrs>=25.1.0",
  "bottleneck>=1.3.4",
  "certifi>=2025.1.31",
  "charset-normalizer>=3.4.1",
  "click>=8.1.8",
  "configobj>=5.0.9",
  "distro>=1.9.0",
  "dnspython>=2.7.0",
  "docker>=7.1.0",
  "docstring-parser>=0.16",
  "email-validator>=2.2.0",
  "fastapi>=0.115.12",
  "fastapi-mcp>=0.3.7",
  "fastapi-versioning>=0.10.0",
  "fastembed>=0.7.1",
  "filelock>=3.17.0",
  "flake8>=3.9.0",
  "fsspec>=2025.2.0",
  "h11>=0.13.0",
  "httpcore>=1.0.7",
  "httpx>=0.28.1",
  "httpx-sse>=0.4.0",
  "huggingface-hub>=0.29.1",
  "idna>=3.10",
  "iniconfig>=2.0.0",
  "jdk4py>=21.0.4.2",
  "jinja2>=3.1.5",
  "jiter>=0.9.0",
  "joblib>=1.4.2",
  "jsonpatch>=1.33",
  "jsonpointer>=3.0.0",
  "langchain-core>=0.3.49",
  "langchain-mcp-adapters>=0.0.6",
  "langchain-openai>=0.3.1",
  "langgraph>=0.2.68",
  "langgraph-checkpoint>=2.1.0",
  "langgraph-prebuilt>=0.5.2",
  "langgraph-sdk>=0.1.70",
  "langsmith>=0.3.22",
  "markdown-it-py>=3.0.0",
  "markupsafe>=3.0.2",
  "mccabe>=0.6.1",
  "mcp>=1.4.1",
  "mdurl>=0.1.2",
  "mpmath>=1.3.0",
  "mypy>=0.910",
  "mypy-extensions>=0.4.4",
  "networkx>=3.4.2",
  "numpy>=1.26.4",
  "openapi-python-client>=0.24.0",
  "orjson>=3.10.16",
  "ormsgpack>=1.10.0",
  "packaging>=24.2",
  "pathspec>=0.12.1",
  "pillow>=11.1.0",
  "platformdirs>=4.3.6",
  "pluggy>=1.5.0",
  "pycodestyle>=2.7.0",
  "pydantic>=2.11.1",
  "pydantic-core",
  "pydantic-settings>=2.8.1",
  "pyflakes>=2.3.1",
  "pygments>=2.19.1",
  "pytest>=8.3.4",
  "pytest-asyncio>=0.25.2",
  "python-dateutil>=2.9.0.post0",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.6",
  "pyyaml>=6.0.2",
  "regex>=2024.11.6",
  "requests>=2.32.3",
  "requests-toolbelt>=1.0.0",
  "rfc3986>=1.5.0",
  "rich>=13.9.4",
  "ruamel-yaml>=0.18.10",
  "ruamel-yaml-clib>=0.2.12",
  "ruff>=0.9.9",
  "safetensors>=0.5.3",
  "scikit-learn>=1.6.1",
  "scipy>=1.15.2",
  "setuptools>=75.8.2",
  "shellingham>=1.5.4",
  "six>=1.17.0",
  "sniffio>=1.3.1",
  "sse-starlette>=2.1.3",
  "starlette>=0.45.3",
  "sympy>=1.13.3",
  "tenacity>=9.0.0",
  "threadpoolctl>=3.5.0",
  "tiktoken>=0.9.0",
  "tokenizers>=0.21.0",
  "toml>=0.10.2",
  "tqdm>=4.67.1",
  "transformers>=4.48.0",
  "typer>=0.15.2",
  "typing-extensions>=4.12.2",
  "urllib3>=1.25.3",
  "uvicorn>=0.23.1",
  "xxhash>=3.5.0",
  "zstandard>=0.23.0",
  "prometheus_client>=0.21.1",
  "opentelemetry-api",
  "opentelemetry-sdk",
  "opentelemetry-exporter-otlp",
  "opentelemetry-instrumentation-fastapi",
  "opentelemetry-instrumentation-requests",
  "faiss-cpu"
]

[project.optional-dependencies]

# Development dependencies
dev = [
  "black>=22.3.0",
  "skillberry-store-sdk @ git+ssh://git@github.ibm.com/skillberry/skillberry-store-sdk.git#subdirectory=skillberry_store_sdk"
]

[tool.uv.sources]
llm-client = { path = "skillberry-common/packages/llm-client", editable = true }

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
sbs = "skillberry_store.main:main"

[tool.setuptools]
include-package-data = true
